First pass at most of the core philosophies
Why is this change needed?
--------------------------
It's important for people to know what makes Expected Behavior tick.

How does it address the issue?
------------------------------
I reformatted the introdution to start with the philosophies and made
a very bad first pass at those files. This is far from ready to be
read. I'm still in the structuring phase.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------