Rewrite "intentionally small"
Why is this change needed?
--------------------------
The first pass had some of the right ideas, but need a lot of
presentation help.

How does it address the issue?
------------------------------
This change is about 70% restructing the existing words so they're in
a better order and 30% adding new words so it more accurately
describes the idea. I think this is a readable draft.

I'm aware it doesn't cover every idea and only skims the surface of
others. I'm trying to balance the overall length of the handbook with
a drive for completeness.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------