A first draft of Guidelines Over Brightlines
Why is this change needed?
--------------------------
It's one of our key philosophies.

How does it address the issue?
------------------------------
This is the first readable draft. I think it covers all the key
concepts, though I'm not sure the tone is right. I'll get some
feedback when I'm done with the overall first draft.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------