First draft of Benefits and How We Work
Why is this change needed?
--------------------------

How does it address the issue?
------------------------------
A lot of this information is copied from other places, including from
- Basecamp
- Chat
- preliminary work done by James to get started on this project

I expanded and updated a lot of the text, so it may not be
recognizable in the current form. Both of these ideas are pretty deep,
so I'm less certain I've covered every base. Nevertheless, I think
it's ready for a review.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------