Try to clarify our Slack communication behavior
Why is this change needed?
--------------------------
Tony suggested in #3 that the wording was confusing. On reread, I agree.

How does it address the issue?
------------------------------
I think the key issue was the absolute "expected" vs. the absolute "indefinite". I softened the expectation by adding a modifier and removed "indefinite". I slightly changed the paragraph structure to add some clarifying comments around being out of slack to focus and being out of slack for vacation.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------