Add more explanation about why we discourage working on vacation
Why is this change needed?
--------------------------
When people choose to work on vacation, they often have good intentions and don't realize the harm they're causing.

How does it address the issue?
------------------------------
I more explicitly stated that working on vacation can be bad for the people that arne't on vacation and added some examples of the kind negative of situations that working on vacation can produce.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------